3-08-22

+++ app/Http/Controllers/AreaController.php
@@ -6,6 +6,8 @@ use App\Models\Area;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
 use App\Http\Request\AreaStoreRequest;
+use Illuminate\Support\Facades\Gate;
+
 
 class AreaController extends Controller
 {
@@ -16,6 +18,7 @@ class AreaController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('area.index'), 403);
         /*$datos['areas'] = Area::paginate(10);
         return view('area.index', $datos);*/
         $areas = Area::paginate(10);
@@ -30,6 +33,7 @@ class AreaController extends Controller
     public function create()
     {
         //
+        abort_if(Gate::denies('area.create'), 403);
         return view('area.create');
     }
 
@@ -86,6 +90,7 @@ class AreaController extends Controller
     public function edit($id)
     {
         //
+        abort_if(Gate::denies('area.update'), 403);
         $area = Area::findOrFail($id);
         return view('area.edite', compact('area'));
     }
@@ -132,6 +137,7 @@ class AreaController extends Controller
     public function destroy($id)
     {
         //
+        abort_if(Gate::denies('area.destroy'), 403);
         Area::destroy($id);
         return redirect('area')->with('info','El rol se elimino con éxito');
     }
